Fix 4px bottom in Firefox causing a scroll bar
Image was defaulting to display: inline and this is beyond me but I made it display: block and now it isn't.
Everything else seems to be fine in both Firefox and Chrome.

I genuinely don't know why this went wrong.
If I ever investigate for some reason: setting background-color: red; on the imageContainer turned it red, so maybe the container was too big for some reason.
Oh yeah, and the imageContainer was 4px taller than the surrounding topContainer, body and html tags. WHY