Fix awful page width on narrow displays; revert scrollable image
Some flex attrs were key to keeping the page width to 100vw.
Removing them enabled scrolling the image, so that's gone again and I'm
sad to see it go, but at least the site looks OK again on half-windows and
mobile devices.

Safari on iOS is pulling some bullshit by positioning the footer
at the bottom when initially drawn and keeping it there when you scroll down,
but fuck Safari on iOS. Firefox Focus on iOS does it properly, interestingly.