Header/footer position: s/absolute/fixed
Not 100% sure on the difference but it works better on mobile devices
this way without appearing to compromise desktop.

https://developer.mozilla.org/en-US/docs/Web/CSS/position

Either it has a special relationship with the viewport (not convinced)
or its stacking context behaviour is better for keeping it attached to
the bottom of the viewport (even less likely).

Rotating my phone still messes with the zoom level so I might have
a go at that.